+++ probe-rs-target/src/chip.rs
@@ -1,6 +1,16 @@
 use super::memory::MemoryRegion;
 use crate::{serialize::hex_option, CoreType};
 use serde::{Deserialize, Serialize};
+
+/// Represents a DAP scan chain element.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub struct ScanChainElement {
+    /// Unique name of the DAP
+    pub name: Option<String>,
+    /// Specifies the IR length of the DAP (default value: 4).
+    pub ir_len: Option<u8>,
+}
+
 /// A single chip variant.
 ///
 /// This describes an exact chip variant, including the cores, flash and memory size. For example,
@@ -43,6 +53,11 @@ pub struct Chip {
     /// executable image that includes the `_SEGGER_RTT` symbol pointing
     /// to the exact address of the RTT header.
     pub rtt_scan_ranges: Option<Vec<std::ops::Range<u64>>>,
+    /// Describes the scan chain
+    ///
+    /// ref: `<https://open-cmsis-pack.github.io/Open-CMSIS-Pack-Spec/main/html/sdf_pg.html#sdf_element_scanchain>`
+    #[serde(default)]
+    pub scan_chain: Option<Vec<ScanChainElement>>,
 }
 
 impl Chip {
@@ -61,6 +76,7 @@ impl Chip {
             memory_map: vec![],
             flash_algorithms: vec![],
             rtt_scan_ranges: None,
+            scan_chain: Some(vec![]),
         }
     }
 }
@@ -109,3 +125,16 @@ pub struct ArmCoreAccessOptions {
 /// The data required to access a Risc-V core
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RiscvCoreAccessOptions {}
+
+/// Helper function that interates the scan chain and returns a vector of all of
+/// the ir_lengths of the scan chain elements.
+/// If an element does not contain an ir_length, the default value of 4 is used.
+/// The first element of the vector is the first element of the scan chain.
+pub fn get_ir_lengths(scan_chain: &Vec<ScanChainElement>) -> Vec<u8> {
+    let mut ir_lengths = Vec::new();
+    for element in scan_chain {
+        let ir_len = element.ir_len.unwrap_or(4);
+        ir_lengths.push(ir_len);
+    }
+    ir_lengths
+}