+++ probe-rs/src/config/target.rs
@@ -1,6 +1,6 @@
-use probe_rs_target::{Architecture, ChipFamily, MemoryRange};
-
-use super::{Core, MemoryRegion, RawFlashAlgorithm, RegistryError, TargetDescriptionSource};
+use super::{
+    Core, MemoryRegion, RawFlashAlgorithm, RegistryError, ScanChainElement, TargetDescriptionSource,
+};
 use crate::architecture::arm::{
     ap::MemoryAp,
     sequences::{
@@ -21,6 +21,7 @@ use crate::architecture::arm::{
 use crate::architecture::riscv::sequences::{esp32c3::ESP32C3, esp32c6::ESP32C6};
 use crate::architecture::riscv::sequences::{DefaultRiscvSequence, RiscvDebugSequence};
 use crate::flashing::FlashLoader;
+use probe_rs_target::{Architecture, ChipFamily, MemoryRange};
 use std::sync::Arc;
 
 use crate::architecture::arm::sequences::DefaultArmSequence;
@@ -45,6 +46,12 @@ pub struct Target {
     /// Each region must be enclosed in exactly one RAM region from
     /// `memory_map`.
     pub rtt_scan_regions: Vec<std::ops::Range<u64>>,
+    /// The Description of the scan chain
+    ///
+    /// The scan chain can be parsed from the CMSIS-SDF file, or specified
+    /// manually in the target.yaml file. It is used by some probes to determine
+    /// the number devices in the scan chain and their ir lengths.
+    pub scan_chain: Option<Vec<ScanChainElement>>,
 }
 
 impl std::fmt::Debug for Target {
@@ -215,6 +222,7 @@ impl Target {
             memory_map: chip.memory_map.clone(),
             debug_sequence,
             rtt_scan_regions,
+            scan_chain: chip.scan_chain.clone(),
         })
     }
 
@@ -320,9 +328,9 @@ impl From<Target> for TargetSelector {
     }
 }
 
-/// This is the type to denote a general debug sequence.  
-/// It can differentiate between ARM and RISC-V for now.  
-/// Currently, only the ARM variant does something sensible;  
+/// This is the type to denote a general debug sequence.
+/// It can differentiate between ARM and RISC-V for now.
+/// Currently, only the ARM variant does something sensible;
 /// RISC-V will be ignored when encountered.
 #[derive(Clone)]
 pub enum DebugSequence {