+++ probe-rs/src/probe.rs
@@ -30,6 +30,7 @@ use crate::{
     Permissions,
 };
 use jlink::list_jlink_devices;
+use probe_rs_target::ScanChainElement;
 use std::{convert::TryFrom, fmt};
 
 /// Used to log warnings when the measured target voltage is
@@ -420,6 +421,20 @@ impl Probe {
         }
     }
 
+    /// Configure the scan chain to use for the attached target.
+    ///
+    /// See [`DebugProbe::set_scan_chain`] for more information and usage
+    pub fn set_scan_chain(
+        &mut self,
+        scan_chain: Vec<ScanChainElement>,
+    ) -> Result<(), DebugProbeError> {
+        if self.attached {
+            self.inner.set_scan_chain(scan_chain)
+        } else {
+            Err(DebugProbeError::Attached)
+        }
+    }
+
     /// Get the currently used maximum speed for the debug protocol in kHz.
     ///
     /// Not all probes report which speed is used, meaning this value is not
@@ -538,6 +553,23 @@ pub trait DebugProbe: Send + fmt::Debug {
     /// `DebugProbeError::UnsupportedSpeed` will be returned.
     ///
     fn set_speed(&mut self, speed_khz: u32) -> Result<u32, DebugProbeError>;
+
+    /// Set the JTAG scan chain information for the target under debug.
+    ///
+    /// This allows the probe to know which TAPs are in the scan chain and their
+    /// position and IR lengths.
+    ///
+    /// If the scan chain is provided, and the selected protocol is JTAG, the
+    /// probe will automatically configure the JTAG interface to match the
+    /// scan chain configuration without trying to deteremine the chain at
+    /// runtime.
+    ///
+    /// This is called by the `Session` when attaching to a target.
+    /// So this does not need to be called manually, unless you want to
+    /// modify the scan chain. You must be attached to a target to set the
+    /// scan_chain since the scan chain only applys to the attached target.
+    ///
+    fn set_scan_chain(&mut self, scan_chain: Vec<ScanChainElement>) -> Result<(), DebugProbeError>;
 
     /// Attach to the chip.
     ///